Changed ints to floats in all graphic items

+++ app/src/main/java/photoshopdroid/mobile/uqac/ca/photoshopdroid/BoundingBox.java
@@ -5,10 +5,10 @@ public class BoundingBox {
     private Point2D inferiorRightCorner;
 
     public BoundingBox(Point2D corner1, Point2D corner2){
-        int x1 = corner1.getPosX();
-        int y1 = corner1.getPosY();
-        int x2 = corner2.getPosX();
-        int y2 = corner2.getPosY();
+        float x1 = corner1.getPosX();
+        float y1 = corner1.getPosY();
+        float x2 = corner2.getPosX();
+        float y2 = corner2.getPosY();
 
         superiorLeftCorner.setPosX(x1 < x2 ? x1 : x2);
         superiorLeftCorner.setPosY(y1 > y2 ? y1 : y2);
